feat: Timeout deployments that are taking too long (#20)
Add a 2 minute timer to the deployment / provisioning while loop.

The server can get into a state, whereby it expects a docker image with
tag revision `N`, and the CLI is already pushing revision `N+1` -- this
could happen for example, if the server failed to reconcile a project in
the `reconcile_project` work queue, but proceeds with a
`reconcile_deploy` work queue item.

In this situation, Cursur will hang indefinitely. We should probably fix
the root problem, but this will at least fail the client after some
time. The end-user can also attempt to delete the project after a
deployment fails and re-create it.